WiFi provisioning started and working! still some rough edges that needs to be cleaned!! -> wrong password: DONE but maybe still little primitive => THIS IS OK -> provisioning after button press (factory reset): DONE -> LED on board will blink when provisioning and stay ON when provisioning is successful.
HTTP connection implemented. There is now a connection possible between the server and the esp32. Over this connection we are able to send data in different kind of request!
DONE: variable content for HTTP message
DONE: put data in JSON!
DONE: make program better for working in overall scenarios.
DONE: being able to add data to the database!!!

Shared data structure:
DONE: implementation of simple data structure with a primitive reader and writer task
TODO: Optimization of the task

SPI configuration started but not finished!

+++ main/http_handler.c
@@ -16,7 +16,7 @@ _Noreturn void http_task()
             if(xQueueReceive(buffer, &data, (TickType_t) 10))
             {
                 int id = data.sensorId;
-                double value = data.value;
+                int value = data.value;
                 http_post_request(id, value);
             }
         }
@@ -70,10 +70,10 @@ esp_http_client_handle_t http_init_connection()
     return client;
 }
 
-void http_post_request(int id, double value)
+void http_post_request(int id, int value)
 {
     char *query = NULL;
-    asprintf(&query, "{\"sensorId\":\"%d\",\"value\":\"%f\"}", id, value); //this is a way of sending JSON.
+    asprintf(&query, "{\"sensorId\":\"%d\",\"value\":\"%d\"}", id, value); //this is a way of sending JSON.
 
     esp_http_client_config_t http_config = {
             .url = DB_URL,
